+++ src/ladb_opencutlist/ruby/constants.rb
@@ -2,6 +2,6 @@ module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
     EXTENSION_VERSION = '1.7.1-dev'.freeze
-    EXTENSION_BUILD = '201906150647'.freeze
+    EXTENSION_BUILD = '201906151524'.freeze
 
 end